+++ v3/crates/compatibility/src/compatibility_date.rs
@@ -173,5 +173,9 @@ pub fn get_compatibility_date_for_flag(flag: Flag) -> Option<CompatibilityDate> 
         | Flag::RequireJwtAudienceValidationIfAudClaimPresent => {
             Some(new_compatibility_date(2025, 3, 11))
         }
+        Flag::DisallowProcedureCommandRelationships => {
+            //TODO: Should be set on next release
+            None
+        }
     }
 }